feat: 4. Data Logging - done

+++ UserManagement.Web/Controllers/UsersController.cs
@@ -1,5 +1,5 @@
 ﻿using System.Linq;
-using System.Security.Cryptography.X509Certificates;
+using Microsoft.Extensions.Logging;
 using UserManagement.Models;
 using UserManagement.Services.Domain.Interfaces;
 using UserManagement.Web.Models.Users;
@@ -10,7 +10,12 @@ namespace UserManagement.WebMS.Controllers;
 public class UsersController : Controller
 {
     private readonly IUserService _userService;
-    public UsersController(IUserService userService) => _userService = userService;
+    private readonly ILogger<UsersController> _logger;
+    public UsersController(IUserService userService, ILogger<UsersController> logger)
+    {
+        _userService = userService;
+        _logger = logger;
+    }
 
     [HttpGet]
     public ViewResult List(bool? filterIsActive)
@@ -100,6 +105,11 @@ public class UsersController : Controller
                 return NotFound();
             }
 
+            //Original users information before edit
+            _logger.LogInformation(new EventId((int)user.Id, "User edit event"),
+            ",{UserId}, {OriginalForename}, {OriginalSurname}, {DateOfBirth:dd/MM/yyyy}, {OriginalEmail}, {OriginalIsActive}, Before Edit",
+            user.Id, user.Forename, user.Surname, user.DateOfBirth, user.Email, user.IsActive);
+
             user.DateOfBirth = model.DateOfBirth;
             user.Email = model.Email ?? user.Email;
             user.IsActive = model.IsActive;
@@ -107,6 +117,11 @@ public class UsersController : Controller
             user.Surname = model.Surname ?? user.Surname;
 
             _userService.UpdateUser(user);
+
+            //New users information after edit
+            _logger.LogInformation(new EventId((int)user.Id, "User edit event"),
+            ",{UserId}, {Forename}, {Surname}, {DateOfBirth:dd/MM/yyyy}, {Email}, {IsActive}, After Edit",
+            user.Id, user.Forename, user.Surname, user.DateOfBirth, user.Email, user.IsActive);
 
             return RedirectToAction("ViewUserDetails", new { userId = user.Id });
         }
@@ -143,6 +158,10 @@ public class UsersController : Controller
 
             _userService.AddUser(newUser);
 
+            _logger.LogInformation(new EventId((int)newUser.Id, "User add event"),
+                ",{UserId}, {Forename}, {Surname}, {DateOfBirth:dd/MM/yyyy}, {Email}, {IsActive}, Add",
+                newUser.Id, newUser.Forename, newUser.Surname, newUser.DateOfBirth, newUser.Email, newUser.IsActive);
+            
             return RedirectToAction("List");
         }
 
@@ -153,7 +172,19 @@ public class UsersController : Controller
     [Route("delete")]
     public IActionResult Delete(long userId)
     {
+        var deletedUser = _userService.GetUserById(userId);
+
+        if (deletedUser == null)
+        {
+            return NotFound();
+        }
+
         _userService.DeleteUser(userId);
+
+        _logger.LogInformation(new EventId((int)userId, "User delete event"),
+        ",{UserId}, {Forename}, {Surname}, {DateOfBirth:dd/MM/yyyy}, {Email}, {IsActive}, Delete",
+        deletedUser.Id, deletedUser.Forename, deletedUser.Surname, deletedUser.DateOfBirth, deletedUser.Email, deletedUser.IsActive);
+
         return RedirectToAction("List");
     }
 }
